requirements.txt
Flask
clingo
flask-jwt-extended
mysql-connector-python
flask-cors